Refresh the session key

Refresh the session key, invalidating all keys created by key_openssl, keypair_openssl, key_sodium and keypair_sodium.

session_key_refresh()

Details

Running this function will invalidate all keys loaded with the above functions. It should not be needed very often.

Examples

# Be careful - if you run this then all keys loaded from file will
# no longer work until reloaded
if (FALSE) {
  cyphr::session_key_refresh()
}