Encrypted data administration

Encrypted data administration; functions for setting up, adding users, etc.

data_admin_init(path_data, path_user = NULL, quiet = FALSE)

data_admin_authorise(path_data, hash = NULL, path_user = NULL,
  yes = FALSE, quiet = FALSE)

data_admin_list_requests(path_data)

data_admin_list_keys(path_data)

Arguments

path_data
Path to the data set. We will store a bunch of things in a hidden directory within this path.

path_user
Path to the directory with your ssh key. Usually this can be omitted.

quiet
Suppress printing of informative messages.

hash
A vector of hashes to add. If provided, each hash can be the binary or string representation of the hash to add. Or omit to add each request.

yes
Skip the confirmation prompt? If any request is declined then the function will throw an error on exit.

Details

data_admin_init initialises the system; it will create a data key if it does not exist and authorise you. If it already exists and you do not have access it will throw an error.

data_admin_authorise authorises a key by creating a key to the data that the user can use in conjunction with their personal key.

data_admin_list_requests lists current requests.

data_admin_list_keys lists known keys that can access the data. Note that this is not secure; keys not listed here may still be able to access the data (if a key was authorised and moved elsewhere for example). Conversely, if the user has deleted or changed their key they will not be able to access the data despite the key being listed here.

See also

Examples

# The workflow here does not really lend itself to an example,
# please see the vignette instead.

# First we need a set of user ssh keys.  In a non example
# environment your personal ssh keys will probably work well, but
# hopefully they are password protected so cannot be used in
# examples.  The password = FALSE argument is only for testing,
# and should not be used for data that you care about.
path_ssh_key <- tempfile()
cyphr::ssh_keygen(path_ssh_key, password = FALSE)

# Initialise the data directory, using this key path.  Ordinarily
# the path_user argument would not be needed because we would be
# using your user ssh keys:
path_data <- tempfile()
dir.create(path_data, FALSE, TRUE)
cyphr::data_admin_init(path_data, path_user = path_ssh_key)

#> Generating data key

#> Authorising ourselves

#> Adding key b5:0b:1b:5b:ac:14:d2:75:12:2c:0f:6c:b8:7e:02:a3
#>   user: rich
#>   host: wpia-dide136.dide.ic.ac.uk
#>   date: 2019-03-25 13:20:28

#> Verifying

# Now you can get the data key
key <- cyphr::data_key(path_data, path_user = path_ssh_key)

# And encrypt things with it
cyphr::encrypt_string("hello", key)

#>  [1] 94 ce 68 71 44 92 59 2e fc af ff f2 ef 8b b7 9e b3 68 a8 1a c0 50 0e 42 d6
#> [26] 99 15 c9 c3 e1 67 4c 9e e3 bf dc ce a9 0b 89 73 d7 fa 5b 58

# See the vignette for more details.  This is not the best medium
# to explore this.

# Cleanup
unlink(path_ssh_key, recursive = TRUE)
unlink(path_data, recursive = TRUE)