User commands

User commands

data_request_access(path_data, path_user = NULL, quiet = FALSE)

data_key(path_data, path_user = NULL, test = TRUE, quiet = FALSE)

Arguments

path_data
Path to the data

path_user
Path to the directory with your user key. Usually this can be omitted. Use the cyphr.user.path global option (i.e., via options()) to set this more conveniently.

quiet
Suppress printing of informative messages.

test
Test that the encryption is working? (Recommended)

Examples

# The workflow here does not really lend itself to an example,
# please see the vignette.

# Suppose that Alice has created a data directory:
path_alice <- tempfile()
cyphr::ssh_keygen(path_alice, password = FALSE)
path_data <- tempfile()
dir.create(path_data, FALSE, TRUE)
cyphr::data_admin_init(path_data, path_user = path_alice)

#> Generating data key

#> Authorising ourselves

#> Adding key 09:20:93:ee:2f:ee:59:9e:94:a9:7d:bc:d2:25:e2:fa
#>   user: rich
#>   host: wpia-dide136.dide.ic.ac.uk
#>   date: 2019-03-25 13:20:28

#> Verifying

# If Bob can also write to the data directory (e.g., it is a
# shared git repo, on a shared drive, etc), then he can request
# access
path_bob <- tempfile()
cyphr::ssh_keygen(path_bob, password = FALSE)
hash <- cyphr::data_request_access(path_data, path_user = path_bob)

#> A request has been added

#> Email someone with access to add you.

#> 	hash: ed:eb:a2:f8:23:c9:1d:f8:17:58:b2:8b:d9:f5:a3:50

# Alice can authorise Bob
cyphr::data_admin_authorise(path_data, path_user = path_alice, yes = TRUE)

#> There is 1 request for access

#> Adding key ed:eb:a2:f8:23:c9:1d:f8:17:58:b2:8b:d9:f5:a3:50
#>   user: rich
#>   host: wpia-dide136.dide.ic.ac.uk
#>   date: 2019-03-25 13:20:28

#> Added 1 key

# After which Bob can get the data key
cyphr::data_key(path_data, path_user = path_bob)

#> <cyphr_key: sodium>

# See the vignette for more details.  This is not the best medium
# to explore this.

# Cleanup
unlink(path_alice, recursive = TRUE)
unlink(path_bob, recursive = TRUE)
unlink(path_data, recursive = TRUE)