Encrypt and decrypt data and other things

Encrypt and decrypt raw data, objects, strings and files. The core functions here are encrypt_data and decrypt_data which take raw data and decrypt it, writing either to file or returning a raw vector. The other functions encrypt and decrypt arbitrary R objects (encrypt_object, decrypt_object), strings (encrypt_string, decrypt_string) and files (encrypt_file, decrypt_file).

encrypt_data(data, key, dest = NULL)

encrypt_object(object, key, dest = NULL, rds_version = NULL)

encrypt_string(string, key, dest = NULL)

encrypt_file(path, key, dest = NULL)

decrypt_data(data, key, dest = NULL)

decrypt_object(data, key)

decrypt_string(data, key)

decrypt_file(path, key, dest = NULL)

Arguments

data
(for encrypt_data, decrypt_data, decrypt_object, decrypt_string) a raw vector with the data to be encrypted or decrypted. For the decryption functions this must be data derived by encrypting something or you will get an error.

key
A cyphr_key or cyphr_key object describing the encryption approach to use.

dest
The destination filename for the encrypted or decrypted data, or NULL to return a raw vector. This is not used by decrypt_object or decrypt_string which always return an object or string.

object
(for encrypt_object) an arbitrary R object to encrypt. It will be serialised to raw first (see serialize).

rds_version
RDS serialisation version to use (see serialize. The default in R version 3.3 and below is version 2 - in the R 3.4 series version 3 was introduced and is becoming the default. Version 3 format serialisation is not understood by older versions so if you need to exchange data with older R versions, you will need to use rds_version = 2. The default argument here (NULL) will ensure the same serialisation is used as R would use by default.

string
(for encrypt_string) a scalar character vector to encrypt. It will be converted to raw first with charToRaw.

path
(for encrypt_file) the name of a file to encrypt. It will first be read into R as binary (see readBin).

Examples

key <- key_sodium(sodium::keygen())
# Some super secret data we want to encrypt:
x <- runif(10)
# Convert the data into a raw vector:
data <- serialize(x, NULL)
data

#>   [1] 58 0a 00 00 00 03 00 04 00 03 00 03 05 00 00 00 00 05 55 54 46 2d 38 00 00
#>  [26] 00 0e 00 00 00 0a 3f b4 ac 0a 80 00 00 00 3f ea b2 db 32 a0 00 00 3f e3 39
#>  [51] 6e e4 e0 00 00 3f c4 1f 67 fd 80 00 00 3f 7e 4e e0 60 00 00 00 3f dd d9 64
#>  [76] 1c 80 00 00 3f df db 95 b1 40 00 00 3f d2 8b 8b e9 c0 00 00 3f e7 73 c4 ec
#> [101] c0 00 00 3f e8 b8 7f 08 40 00 00

# Encrypt the data; without the key above we will never be able to
# decrypt this.
data_enc <- encrypt_data(data, key)
data_enc

#>   [1] a0 ef 45 31 eb e0 c8 da b8 42 4b b6 c0 cc 1a 1f 7e 30 83 49 87 b1 e5 84 29
#>  [26] 46 73 3c c2 bc 90 94 16 37 4a 61 da d8 02 03 ae 64 4c 27 fe f0 03 81 5d 44
#>  [51] 90 50 44 c9 a4 54 f8 c8 9f d7 ee 8e 8c b4 bb 9a eb fc 56 d5 25 5d bc 46 2a
#>  [76] 52 ba b3 25 b0 3a 0f 68 2d 07 79 62 5c c5 6c 4d 9c 4a 6f 44 7c 8e 2e 1e 79
#> [101] a5 c8 e2 4c c4 1c d3 7e e4 ae 01 11 80 82 ad 87 fa 36 43 60 42 7c 3a 96 54
#> [126] d0 35 89 09 a1 7d db b9 12 3d bd 6e 89 8c 95 ca 65 03 6a 95 94 4a 7c 50 76
#> [151] 33

# Our random numbers:
unserialize(decrypt_data(data_enc, key))

#>  [1] 0.080750138 0.834333037 0.600760886 0.157208442 0.007399441 0.466393497
#>  [7] 0.497777389 0.289767245 0.732881987 0.772521511

# Same as the never-encrypted version:
x

#>  [1] 0.080750138 0.834333037 0.600760886 0.157208442 0.007399441 0.466393497
#>  [7] 0.497777389 0.289767245 0.732881987 0.772521511

# This can be achieved more easily using `encrypt_object`:
data_enc <- encrypt_object(x, key)
identical(decrypt_object(data_enc, key), x)

#> [1] TRUE

# Encrypt strings easily:
str_enc <- encrypt_string("secret message", key)
str_enc

#>  [1] b5 e7 ab af 85 3c 99 ff 66 2f d2 d5 99 26 ec 5b 2d 5f 81 87 f1 2c 90 93 8d
#> [26] 5f 79 2c a0 58 44 0a 52 e9 a7 32 79 a7 8d 1b a4 de 24 3d 27 57 41 9d bf 4d
#> [51] 4e 4a 46 8d

decrypt_string(str_enc, key)

#> [1] "secret message"